+++ src/lib/types/tx/index.ts
@@ -1,3 +1,4 @@
 export * from "./msg";
 export * from "./model";
 export * from "./transaction";
+export * from "./common";